pyproject.toml
[tool.poetry]
name = "zenmoney-bot"
version = "0.1.0"
description = ""
authors = ["Vlad Pastushenko <iam@vladpi.me>"]

[tool.poetry.dependencies]
python = "3.9.2"
fastapi = "^0.68.1"
uvicorn = {version = "^0.15.0", extras = ["standart"]}
python-dotenv = "^0.19.0"
colorlog = "^6.4.1"
httpx = "^0.19.0"
aiogram = "^2.15"
fastapi-security-telegram-webhook = "^0.2.0"
databases = {version = "^0.5.2", extras = ["postgresql"]}
alembic = "^1.7.3"
psycopg2 = "^2.9.1"
pycryptodome = "^3.10.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^21.8b0"
isort = "^5.9.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
